tr1/output: move textures to trx
