+++ src/tr1/game/game/game_draw.c
@@ -2,7 +2,6 @@
 
 #include "game/lara.h"
 #include "game/output/scene_compositor.h"
-#include "game/output/textures.h"
 #include "game/overlay.h"
 #include "game/room_draw.h"
 
@@ -10,6 +9,7 @@
 #include <libtrx/game/camera.h>
 #include <libtrx/game/interpolation.h>
 #include <libtrx/game/lara/hair.h>
+#include <libtrx/game/output/textures.h>
 
 void Game_Draw(bool draw_overlay)
 {
